+++ pcmwav.h
@@ -21,33 +21,33 @@
 #pragma pack(push, 1)
 
 typedef struct {
-	unsigned long	ChunkID;		// 'RIFF'
-	unsigned long	ChunkSize;		
-	unsigned long	Format;			// 'WAVE'
+	unsigned int	ChunkID;		// 'RIFF'
+	unsigned int	ChunkSize;
+	unsigned int	Format;			// 'WAVE'
 } RIFFhdr;
 
 typedef struct {
-	unsigned long	Subchunk1Size;
+	unsigned int	Subchunk1Size;
 	unsigned short	AudioFormat;	// PCM = 1
 	unsigned short	NumChannels;
-	unsigned long	SampleRate;
-	unsigned long	ByteRate;
+	unsigned int	SampleRate;
+	unsigned int	ByteRate;
 	unsigned short	BlockAlign;
 	unsigned short	BitsPerSample;
 } fmt_sub;
 
 typedef struct {
 	unsigned short	nchannels;		// number of channels
-	unsigned long	samplerate;		// sampling rate (e.g. 44100)
-	unsigned long	bitspersample;	// bits per sample (8, 16)
-	unsigned long	ndatabytes;		// number of data bytes in wave file
+	unsigned int	samplerate;		// sampling rate (e.g. 44100)
+	unsigned int	bitspersample;	// bits per sample (8, 16)
+	unsigned int	ndatabytes;		// number of data bytes in wave file
 
 	// private variables
 	FILE			*winfile;		// file handle
-	unsigned long	formatpos;
-	unsigned long	datapos;
-	unsigned long	datasizepos;
-	unsigned long	filesize;
+	unsigned int	formatpos;
+	unsigned int	datapos;
+	unsigned int	datasizepos;
+	unsigned int	filesize;
 } pcmwavfile;
 
 #pragma pack(pop)
@@ -60,16 +60,16 @@ extern char pcmwav_error[];	// On error: contains a string that describes the er
 int pcmwav_open(const char *fname, const char *access, pcmwavfile *opwf);
 
 // Reads len data bytes (not samples!) into buf
-int pcmwav_read(pcmwavfile *pwf, void *buf, unsigned long len);
+int pcmwav_read(pcmwavfile *pwf, void *buf, size_t len);
 
 // Writes len data bytes from buf
-int pcmwav_write(pcmwavfile *pwf, void *buf, unsigned long len);
+int pcmwav_write(pcmwavfile *pwf, void *buf, size_t len);
 
 // Rewinds to start of data
 int pcmwav_rewind(pcmwavfile *pwf);
 
 // Seeks +/- pos in file
-int pcmwav_seek(pcmwavfile *pwf, long pos);
+int pcmwav_seek(pcmwavfile *pwf, size_t pos);
 
 // Closes PCM WAV file
 int pcmwav_close(pcmwavfile *pwf);
